Start new jobs on changed autoScheduleFrequency
